Move user requests to device.js. Move user tests to device.coffee.
