+++ lib/cam.js
@@ -919,4 +919,3 @@ require('./ptz')(Cam);
 require('./imaging')(Cam);
 require('./recording')(Cam);
 require('./replay')(Cam);
-require('./user')(Cam);
